refactor(manager): extract camera state enums to state-enums.go (#33)
Split state.go (375 lines) into 2 files:
- state-enums.go: CameraState type + constants + String(), CameraSubstate type + constants + String()
- state.go: SimState struct, Equal() method, SimStateChange struct, handlers/subscription types, simStateDataStruct

No behavior changes — pure file reorganization.
